removed file.ts
